More mouse actions
Left click now moves the worker to the cell the player clicked on assuming
there is a path to that cell which you can take without pushing any boxes.
Right click undoes a single move.  The former default for clicks, i.e. move
in parallel to the axes the player towards a certain cell or not at all,
can now be executed using a click with the middle mouse button.